+++ changes.md
@@ -2,3 +2,5 @@ Changes done:
     - refactored controllers (authController & tripController) to use promise format
     - corrected HTTP methods
     - renamed all 'auth' items to use "user" instead;
+    - added tests for userController & tripController
+    - added Intgration Test for Login Process